Refactor rest (#320)
* reactivate redis queue

* rename variable

* make queue type configurable

* enhance dev setup for redis queue

* lint

* move user_auth

* move user auth

* move base_login

* move map_layer_base

* move renderer_base

* move resource_base

* fix relative imports

* lint
