+++ src/actinia_core/rest/base/resource_base.py
@@ -33,6 +33,8 @@ from flask import make_response, jsonify
 from flask import request, g
 from flask.json import loads as json_loads
 from flask_restful_swagger_2 import Resource
+from actinia_core.rest.base.user_auth import check_user_permissions
+from actinia_core.rest.base.user_auth import create_dummy_user
 from actinia_core.core.common.app import auth
 from actinia_core.core.common.app import flask_api
 from actinia_core.core.common.config import global_config
@@ -42,9 +44,8 @@ from actinia_core.core.resources_logger import ResourceLogger
 from actinia_core.core.resource_data_container import ResourceDataContainer
 from actinia_core.models.response_models import ProcessingResponseModel
 from actinia_core.models.response_models import create_response_from_model, ApiInfoModel
-from .resource_streamer import RequestStreamerResource
-from .user_auth import check_user_permissions, create_dummy_user
-from .resource_management import ResourceManager
+from actinia_core.rest.resource_streamer import RequestStreamerResource
+from actinia_core.rest.resource_management import ResourceManager
 
 __license__ = "GPLv3"
 __author__ = "Sören Gebbert, Anika Weinmann"
