+++ src/actinia_core/rest/download_cache_management.py
@@ -33,15 +33,15 @@ from flask_restful_swagger_2 import swagger
 import pickle
 from actinia_core.rest.ephemeral_processing import EphemeralProcessing
 from actinia_core.rest.persistent_processing import PersistentProcessing
-from actinia_core.rest.resource_base import ResourceBase
+from actinia_core.rest.base.resource_base import ResourceBase
 from actinia_core.core.common.redis_interface import enqueue_job
 from actinia_core.core.common.process_object import Process
 from actinia_core.core.common.exceptions import AsyncProcessError
 from actinia_core.models.response_models import \
     StorageResponseModel, StorageModel, ProcessingResponseModel
 from actinia_core.core.common.api_logger import log_api_call
-from actinia_core.rest.user_auth import very_admin_role
-from actinia_core.rest.user_auth import check_user_permissions
+from actinia_core.rest.base.user_auth import very_admin_role
+from actinia_core.rest.base.user_auth import check_user_permissions
 from actinia_core.core.common.app import auth
 
 __license__ = "GPLv3"
