+++ src/actinia_core/rest/ephemeral_processing_with_export.py
@@ -32,7 +32,7 @@ from flask import jsonify, make_response
 from copy import deepcopy
 from flask_restful_swagger_2 import swagger
 from actinia_core.rest.ephemeral_processing import EphemeralProcessing
-from actinia_core.rest.resource_base import ResourceBase
+from actinia_core.rest.base.resource_base import ResourceBase
 from actinia_core.core.common.redis_interface import enqueue_job
 from actinia_core.core.common.process_object import Process
 from actinia_core.core.common.process_chain import ProcessChainModel
