+++ src/actinia_core/rest/mapset_management.py
@@ -37,13 +37,13 @@ from actinia_api.swagger2.actinia_core.schemas.mapset_management import \
     MapsetLockManagementResponseModel
 
 from actinia_core.rest.persistent_processing import PersistentProcessing
-from actinia_core.rest.resource_base import ResourceBase
+from actinia_core.rest.base.resource_base import ResourceBase
 from actinia_core.core.common.app import auth
 from actinia_core.core.common.api_logger import log_api_call
 from actinia_core.core.common.redis_interface import enqueue_job
 from actinia_core.core.common.exceptions import AsyncProcessError
-from actinia_core.rest.user_auth import check_user_permissions
-from actinia_core.rest.user_auth import very_admin_role
+from actinia_core.rest.base.user_auth import check_user_permissions
+from actinia_core.rest.base.user_auth import very_admin_role
 from actinia_core.models.response_models import ProcessingResponseModel, \
     StringListProcessingResultResponseModel, MapsetInfoResponseModel, \
     RegionModel, ProcessingErrorResponseModel
