+++ src/actinia_core/rest/resource_storage_management.py
@@ -33,7 +33,7 @@ from flask_restful_swagger_2 import swagger
 import pickle
 from .ephemeral_processing import EphemeralProcessing
 from .persistent_processing import PersistentProcessing
-from .resource_base import ResourceBase
+from actinia_core.rest.base.resource_base import ResourceBase
 from actinia_core.core.common.redis_interface import enqueue_job
 from actinia_core.core.common.process_object import Process
 from actinia_core.core.common.exceptions import AsyncProcessError
@@ -43,8 +43,8 @@ from actinia_core.models.response_models import \
 
 from actinia_core.core.common.app import auth
 from actinia_core.core.common.api_logger import log_api_call
-from .user_auth import very_admin_role
-from .user_auth import check_user_permissions
+from actinia_core.rest.base.user_auth import very_admin_role
+from actinia_core.rest.base.user_auth import check_user_permissions
 
 __license__ = "GPLv3"
 __author__ = "Sören Gebbert"
