+++ src/actinia_core/rest/strds_renderer.py
@@ -27,7 +27,8 @@ Raster map renderer
 from flask import jsonify, make_response, Response
 from .ephemeral_processing import EphemeralProcessing
 from actinia_core.core.common.redis_interface import enqueue_job
-from .renderer_base import RendererBaseResource, EphemeralRendererBase
+from actinia_core.rest.base.renderer_base import RendererBaseResource, \
+     EphemeralRendererBase
 
 import os
 from flask_restful_swagger_2 import swagger
