+++ src/actinia_core/rest/user_management.py
@@ -33,7 +33,7 @@ TODO: Implement POST full permission creation
 from flask import jsonify, make_response
 from flask_restful import reqparse
 from flask_restful_swagger_2 import swagger
-from .base_login import LoginBase
+from actinia_core.rest.base.base_login import LoginBase
 from actinia_core.core.common.user import ActiniaUser
 from actinia_core.models.response_models import \
     UserListResponseModel, UserInfoResponseModel, SimpleResponseModel
